refactor: migrate to `esm-env`
