+++ package.json
@@ -58,6 +58,7 @@
 	},
 	"dependencies": {
 		"@prismicio/client": "^7.12.0",
+		"esm-env": "^1.2.1",
 		"vue-router": "^4.5.0"
 	},
 	"devDependencies": {
